requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==4.4.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.26.1
imageio-ffmpeg==0.4.8
jedi==0.18.2
moviepy==1.0.3
numpy==1.24.2
packaging==23.0
parso==0.8.3
Pillow==9.4.0
proglog==0.1.10
pudb==2022.1.3
pycparser==2.21
Pygments==2.14.0
python-telegram-bot==13.8
pytube==12.1.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tqdm==4.65.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
urwid==2.1.2
urwid-readline==0.13
